move the assertions into utils/value.ts
