+++ compiler/src/macros/commands/UnitRadar.ts
@@ -3,12 +3,12 @@ import { MacroFunction } from "..";
 import { LiteralValue, SenseableValue, StoreValue } from "../../values";
 import { validRadarFilters, validRadarSorts } from "./Radar";
 import { CompilerError } from "../../CompilerError";
+import { EMutability } from "../../types";
 import {
   assertIsArrayMacro,
   assertIsObjectMacro,
   assertLiteralOneOf,
-} from "../../assertions";
-import { EMutability } from "../../types";
+} from "../../utils";
 
 export class UnitRadar extends MacroFunction {
   constructor() {
